Add iOS Navigation Tests workflow to GitHub Actions
This commit introduces a new workflow in `ios_e2e_tests.yml` for building and running iOS navigation tests. The workflow is triggered on pull requests to the `main` branch and includes steps for environment setup, including Node.js, .NET, and Docker installation, as well as starting an Appium server and managing an iOS simulator. It also handles restoring and building a MAUI app and running the tests, with an option to upload Appium logs on failure.