Add iOS Navigation Tests workflow to GitHub Actions
This commit introduces a new GitHub Actions workflow named "Build and Run iOS Navigation Tests" in the `ios_navigation_tests.yml` file. The workflow is triggered on pull requests to the `main` branch and runs on macOS 14. It sets up the environment for testing an iOS application by installing Node.js, .NET, Appium, and necessary workloads. The workflow also creates and boots an iOS simulator, restores the MAUI app, builds the code, and executes the iOS navigation tests. Additionally, it uploads Appium logs in case of test failures.